combined alerts and preferences
