--- src/routes/settings/preferences/+layout.svelte
@@ -1,17 +1,0 @@
-<script>
-	import '../../../app.css';
-	import '../../styles.css';
-</script>
-
-<div class="app">
-	
-	<main>
-		<slot></slot> <!-- page.svelte content goes here-->
-	</main>
-
-	<footer>
-	</footer>
-</div>
-
-<style>
-</style>
